refactor: remove default data provider
