--- src/providers/data-provider/index.ts
@@ -1,1 +1,0 @@
-
